requirements.txt
orjson
rich
dynaconf
pydantic[email]
loguru
aiodns
fastapi[standard]
aioquic
hypercorn[h3]
pyjwt
passlib
argon2-cffi
asyncpg
httpx[http2]
httpx-sse
lark
aiomudtelnet @ git+https://github.com/volundmush/aiomudtelnet@main
cython
setuptools